Added endpoints to mealRouter and nutritionRouter

+++ src/services/mealService.ts
@@ -1,8 +1,16 @@
-import { $Enums, Meal, PrismaClient, ServingUnit } from "@prisma/client";
+import { $Enums, PrismaClient, ServingUnit } from "@prisma/client";
 import { UserNotFound } from "../exceptions/UserNotFound";
-import { convert } from "./unitService";
+import { calcNutritionForItem, calcNutritionForMeal } from "./nutritionService";
+import { MealNotFound } from "../exceptions/MealNotFound";
+import { MealNotBelongToUser } from "../exceptions/MealNotBelongToUser";
+import { getUnitByAbbrev } from "./unitService";
+import { UnitNotFound } from "../exceptions/UnitNotFound";
 
 const prisma = new PrismaClient();
+
+export function isMealName(name: string) {
+  return (Object.values($Enums.MealType) as string[]).includes(name);
+}
 
 export async function createMeal(
   username: string,
@@ -32,8 +40,10 @@ export async function addFoodToMeal(
   mealId: string,
   foodId: string,
   amount: number,
-  unit: ServingUnit
+  unitAbbrev: string
 ) {
+  const unit = await getUnitByAbbrev(unitAbbrev);
+  if (!unit) throw new UnitNotFound();
   return await prisma.mealItem.create({
     data: {
       amount,
@@ -56,7 +66,31 @@ export async function addFoodToMeal(
   });
 }
 
-export async function deleteMealItem(mealItemId: string) {
+export async function getMealItemById(mealItemId: string) {
+  return await prisma.mealItem.findUnique({
+    where: {
+      id: mealItemId,
+    },
+    include: {
+      meal: {
+        include: { user: true },
+      },
+      food: true,
+      unit: true,
+    },
+  });
+}
+
+export async function checkMealItemOwnership(
+  username: string,
+  mealItemId: string
+) {
+  const realUsername = (await getMealItemById(mealItemId)).meal.user.username;
+  if (realUsername !== username) throw new MealNotBelongToUser();
+}
+
+export async function deleteMealItem(username: string, mealItemId: string) {
+  await checkMealItemOwnership(username, mealItemId);
   return await prisma.mealItem.delete({
     where: {
       id: mealItemId,
@@ -65,10 +99,14 @@ export async function deleteMealItem(mealItemId: string) {
 }
 
 export async function updateMealItem(
+  username: string,
   mealItemId: string,
   amount: number,
-  unit: ServingUnit
+  unitAbbrev: string
 ) {
+  await checkMealItemOwnership(username, mealItemId);
+  const unit = await getUnitByAbbrev(unitAbbrev);
+  if (!unit) throw new UnitNotFound();
   return await prisma.mealItem.update({
     where: {
       id: mealItemId,
@@ -84,40 +122,42 @@ export async function updateMealItem(
   });
 }
 
-async function calcNutritionForItem(item) {
-  const ratio =
-    (await convert(item.unit.id, item.food.baseServingUnit.id, item.amount)) /
-    item.food.baseServingSize;
-  return {
-    calories: item.food.calories * ratio,
-    protein: item.food.protein * ratio,
-    carbs: item.food.carbs * ratio,
-    fat: item.food.fat * ratio,
-    sugar: item.food.sugar * ratio,
-  };
+export async function getMealById(mealId: string) {
+  const meal = await prisma.meal.findUnique({
+    where: {
+      id: mealId,
+    },
+    include: {
+      items: {
+        include: {
+          food: true,
+          unit: true,
+        },
+      },
+    },
+  });
+  if (!meal) throw new MealNotFound();
+  return await makeMealWithNutrition(meal);
 }
 
-async function calcTotalNutrition(meal) {
-  let result = {
-    calories: 0,
-    protein: 0,
-    carbs: 0,
-    fat: 0,
-    sugar: 0,
+async function makeMealWithNutrition(meal) {
+  return {
+    id: meal.id,
+    name: meal.name,
+    totalNutrition: await calcNutritionForMeal(meal),
+    date: meal.mealDate,
+    items: await Promise.all(
+      meal.items.map(async (item) => {
+        return {
+          nutrition: await calcNutritionForItem(item),
+          id: item.id,
+          amount: item.amount,
+          unit: item.unit,
+          food: item.food,
+        };
+      })
+    ),
   };
-
-  for (const item of meal.items) {
-    const { calories, protein, fat, carbs, sugar } = await calcNutritionForItem(
-      item
-    );
-    result.calories += calories;
-    result.protein += protein;
-    result.carbs += carbs;
-    result.fat += fat;
-    result.sugar += sugar;
-  }
-
-  return result;
 }
 
 export async function getMealsForDay(username: string, date: string) {
@@ -142,25 +182,5 @@ export async function getMealsForDay(username: string, date: string) {
     },
   });
 
-  return await Promise.all(
-    rawResult.map(async (meal) => {
-      return {
-        id: meal.id,
-        name: meal.name,
-        totalNutrition: await calcTotalNutrition(meal),
-        date: meal.mealDate,
-        items: await Promise.all(
-          meal.items.map(async (item) => {
-            return {
-              nutrition: await calcNutritionForItem(item),
-              id: item.id,
-              amount: item.amount,
-              unit: item.unit,
-              food: item.food,
-            };
-          })
-        ),
-      };
-    })
-  );
+  return await Promise.all(rawResult.map(makeMealWithNutrition));
 }
